requirements.txt
psycopg2-binary
django-rest-framework
Django
